// apps/auth-frontend/src/features/auth/register/schemas/register.schema.ts
import { z } from 'zod'

export const registerSchema = z.object({
  name: z.string().min(2, 'Nome deve ter ao menos 2 caracteres'),
  email: z.string().email('E-mail inválido'),
  password: z.string().min(8, 'Senha deve ter ao menos 8 caracteres'),
})

export type RegisterFormValues = z.infer<typeof registerSchema>